Move MCPTool type to tyes file
