+++ packages/server/api/src/app/ai/mcp/tables-tools.ts
@@ -5,7 +5,7 @@ import {
 import { AppSystemProp, system } from '@openops/server-shared';
 import { experimental_createMCPClient as createMCPClient, ToolSet } from 'ai';
 import { openopsTables } from '../../openops-tables';
-import { MCPTool } from './tools-initializer';
+import { MCPTool } from './types';
 
 export async function getTablesTools(): Promise<MCPTool> {
   const { token } = await authenticateDefaultUserInOpenOpsTables();
